@@ -20,11 +20,12 @@
         }
     },
     "require": {
-        "php": "^8.0",
+        "php": "^8.1",
         "react/socket": "^1.12",
         "react/datagram": "^1.9",
-        "graylog2/gelf-php": "^1.5|2.0",
+        "graylog2/gelf-php": "^2.0",
         "symfony/process": "^5.4|^6.0",
+        "ext-sockets": "*",
         "ext-json": "*",
         "ext-zlib": "*"
     },
@@ -35,8 +36,7 @@
         "phpstan/phpstan-symfony": "^1.2",
         "phpunit/phpunit": "^9.5",
         "sempro/phpunit-pretty-print": "^1.4",
-        "symfony/dotenv": "^5.4|^6.0",
-        "ext-sockets": "*"
+        "symfony/dotenv": "^5.4|^6.0"
     },
     "scripts": {
         "tests": "phpunit --coverage-clover build/logs/clover.xml --coverage-xml=build/logs/coverage-xml --log-junit=build/logs/phpunit.junit.xml",
