@@ -6,6 +6,5 @@ use Symfony\Component\Dotenv\Dotenv;
 
 require __DIR__ . '/../vendor/autoload.php';
 
-if (file_exists(dirname(__DIR__).'/.env.local')) {
-    (new Dotenv())->usePutenv(true)->bootEnv(dirname(__DIR__).'/.env.local', 'dev', []);
-}
+(new Dotenv())->usePutenv()->bootEnv(dirname(__DIR__).'/.env', 'dev', []);
+
